Relocate agentic implementation to its own _src/agentic/ subpackage
The orchestrator, prompts, and Lookup utility used to live under
_src/optimization/ and _src/datageneration/ — residue from v1 when
AgentOptimizer subclassed Optimizer. In v2 the orchestrator is
AgenticRun (no Optimizer relation), so the placement was just
co-locating agentic code with unrelated native f3dasm modules.

This commit moves:

  _src/optimization/agent_runtime.py   →  _src/agentic/agent_runtime.py
  _src/optimization/agent_prompts.py   →  _src/agentic/agent_prompts.py
  _src/datageneration/lookup.py        →  _src/agentic/lookup.py
  tests/optimization/test_agent_*.py   →  tests/agentic/test_agent_*.py
  tests/datageneration/test_lookup.py  →  tests/agentic/test_lookup.py

All renames via git mv (history preserved). Public import paths
(`from f3dasm.agentic import ...`) unchanged — only the
implementation imports in src/f3dasm/agentic/__init__.py,
src/f3dasm/agentic/__main__.py, and the test files were
rewritten. The Implementer's f3dasm_primer was updated to
recommend the public `from f3dasm.agentic import
LookupDataGenerator` path.

Native f3dasm directories (_src/optimization/ and
_src/datageneration/) now contain zero agentic code. 138/138
tests pass; ruff clean.
