+++ src/f3dasm/_src/agentic/lookup.py
