+++ tests/agentic/test_agent_prompts.py
@@ -1,4 +1,4 @@
-"""Tests for ``f3dasm._src.optimization.agent_prompts``.
+"""Tests for ``f3dasm._src.agentic.agent_prompts``.
 
 Validates that the four prompt constants shipped with the agentic-f3dasm
 v2 runtime satisfy structural, content, and integration requirements.
@@ -85,7 +85,7 @@ def test_constants_exist_and_are_non_empty_strings():
     ``IMPLEMENTER_RESET_PROMPT_TEMPLATE`` has a lower threshold (300 chars)
     because it is a short wrapper template, not a full system prompt.
     """
-    from f3dasm._src.optimization.agent_prompts import (
+    from f3dasm._src.agentic.agent_prompts import (
         CHECKPOINT_STRATEGIZER_PROMPT,
         IMPLEMENTER_RESET_PROMPT_TEMPLATE,
         IMPLEMENTER_SYSTEM_PROMPT,
@@ -122,7 +122,7 @@ def test_strategizer_xml_sections_appear_exactly_once():
     The runtime relies on these sections to give structure to the prompt;
     duplicate tags would confuse any downstream XML parser.
     """
-    from f3dasm._src.optimization.agent_prompts import (
+    from f3dasm._src.agentic.agent_prompts import (
         STRATEGIZER_SYSTEM_PROMPT,
     )
 
@@ -151,7 +151,7 @@ def test_strategizer_names_all_failure_modes():
     The names are matched case-insensitively to allow stylistic variation
     (e.g. ``ANCHORING BIAS`` vs ``anchoring``).
     """
-    from f3dasm._src.optimization.agent_prompts import (
+    from f3dasm._src.agentic.agent_prompts import (
         STRATEGIZER_SYSTEM_PROMPT,
     )
 
@@ -182,7 +182,7 @@ def test_strategizer_mentions_all_five_tools():
     The tool names are checked as literal substrings; the prompt must
     spell them exactly so the model sees the correct function names.
     """
-    from f3dasm._src.optimization.agent_prompts import (
+    from f3dasm._src.agentic.agent_prompts import (
         STRATEGIZER_SYSTEM_PROMPT,
     )
 
@@ -205,7 +205,7 @@ def test_strategizer_briefing_clarification_ritual():
     A loose proximity check (within 150 chars) is used rather than an
     exact sentence match, to remain robust to minor prose edits.
     """
-    from f3dasm._src.optimization.agent_prompts import (
+    from f3dasm._src.agentic.agent_prompts import (
         STRATEGIZER_SYSTEM_PROMPT,
     )
 
@@ -243,7 +243,7 @@ def test_implementer_xml_sections_appear_exactly_once():
     Includes the f3dasm-specific ``<f3dasm_primer>`` tag that teaches the
     Implementer how to use the framework.
     """
-    from f3dasm._src.optimization.agent_prompts import (
+    from f3dasm._src.agentic.agent_prompts import (
         IMPLEMENTER_SYSTEM_PROMPT,
     )
 
@@ -273,7 +273,7 @@ def test_implementer_f3dasm_primer_references_key_classes():
     Checked as literal substrings (case-sensitive) because the model
     must use the exact class names when generating code.
     """
-    from f3dasm._src.optimization.agent_prompts import (
+    from f3dasm._src.agentic.agent_prompts import (
         IMPLEMENTER_SYSTEM_PROMPT,
     )
 
@@ -303,7 +303,7 @@ def test_implementer_report_format_headings():
     The runtime greps for ``## Report`` to extract the Implementer's
     response; the other headings are checked here as a structural contract.
     """
-    from f3dasm._src.optimization.agent_prompts import (
+    from f3dasm._src.agentic.agent_prompts import (
         IMPLEMENTER_SYSTEM_PROMPT,
     )
 
@@ -333,7 +333,7 @@ def test_implementer_refuses_hypothesis_verification():
     Both ``hypothesis`` and ``refuse`` must appear (case-insensitive) so
     the model is clearly instructed to surface scope violations.
     """
-    from f3dasm._src.optimization.agent_prompts import (
+    from f3dasm._src.agentic.agent_prompts import (
         IMPLEMENTER_SYSTEM_PROMPT,
     )
 
@@ -359,7 +359,7 @@ def test_no_supercompressible_leakage():
     supercompressible metamaterial study would bias the agent toward that
     domain.
     """
-    from f3dasm._src.optimization.agent_prompts import (
+    from f3dasm._src.agentic.agent_prompts import (
         CHECKPOINT_STRATEGIZER_PROMPT,
         IMPLEMENTER_RESET_PROMPT_TEMPLATE,
         IMPLEMENTER_SYSTEM_PROMPT,
@@ -403,7 +403,7 @@ def test_checkpoint_prompt_section_headings():
     These headings are case-sensitive because the Strategizer model is
     instructed to reproduce them verbatim in its checkpoint report.
     """
-    from f3dasm._src.optimization.agent_prompts import (
+    from f3dasm._src.agentic.agent_prompts import (
         CHECKPOINT_STRATEGIZER_PROMPT,
     )
 
@@ -431,7 +431,7 @@ def test_checkpoint_prompt_forbids_new_hypotheses():
     -----
     Either of two acceptable phrasings is checked case-insensitively.
     """
-    from f3dasm._src.optimization.agent_prompts import (
+    from f3dasm._src.agentic.agent_prompts import (
         CHECKPOINT_STRATEGIZER_PROMPT,
     )
 
@@ -459,7 +459,7 @@ def test_reset_template_has_exactly_one_placeholder():
     A ``KeyError`` from ``.format()`` would indicate a stray placeholder
     that the runtime does not supply.
     """
-    from f3dasm._src.optimization.agent_prompts import (
+    from f3dasm._src.agentic.agent_prompts import (
         IMPLEMENTER_RESET_PROMPT_TEMPLATE,
     )
 
@@ -494,7 +494,7 @@ def test_reset_template_structure_after_formatting():
     ``Strategizer`` and ``Task`` are checked as loose anchors that orient
     the fresh Implementer session to its role and workflow.
     """
-    from f3dasm._src.optimization.agent_prompts import (
+    from f3dasm._src.agentic.agent_prompts import (
         IMPLEMENTER_RESET_PROMPT_TEMPLATE,
     )
 
@@ -530,7 +530,7 @@ def test_runtime_imports_all_four_constants():
     import pathlib
 
     runtime_path = pathlib.Path(
-        "src/f3dasm/_src/optimization/agent_runtime.py"
+        "src/f3dasm/_src/agentic/agent_runtime.py"
     )
     # Resolve relative to the repo root (two parents above tests/).
     if not runtime_path.is_absolute():
@@ -563,7 +563,7 @@ def test_runtime_imports_all_four_constants():
 
 def test_strategizer_hypothesis_log_tag_appears_once():
     """STRATEGIZER_SYSTEM_PROMPT has exactly one <hypothesis_log> pair."""
-    from f3dasm._src.optimization.agent_prompts import (
+    from f3dasm._src.agentic.agent_prompts import (
         STRATEGIZER_SYSTEM_PROMPT,
     )
 
@@ -576,7 +576,7 @@ def test_strategizer_hypothesis_log_tag_appears_once():
 
 def test_strategizer_hypothesis_log_content():
     """hypothesis_log section mentions all required fields."""
-    from f3dasm._src.optimization.agent_prompts import (
+    from f3dasm._src.agentic.agent_prompts import (
         STRATEGIZER_SYSTEM_PROMPT,
     )
 
@@ -601,7 +601,7 @@ def test_strategizer_hypothesis_log_content():
 
 def test_strategizer_on_error_tag_appears_once():
     """STRATEGIZER_SYSTEM_PROMPT has exactly one <on_error> pair."""
-    from f3dasm._src.optimization.agent_prompts import (
+    from f3dasm._src.agentic.agent_prompts import (
         STRATEGIZER_SYSTEM_PROMPT,
     )
 
@@ -614,7 +614,7 @@ def test_strategizer_on_error_tag_appears_once():
 
 def test_strategizer_on_error_content():
     """on_error mentions REFLECT: and forbids identical re-delegation."""
-    from f3dasm._src.optimization.agent_prompts import (
+    from f3dasm._src.agentic.agent_prompts import (
         STRATEGIZER_SYSTEM_PROMPT,
     )
 
@@ -638,7 +638,7 @@ def test_strategizer_on_error_content():
 
 def test_checkpoint_prompt_contains_comment_log():
     """CHECKPOINT_STRATEGIZER_PROMPT contains the ### Comment log section."""
-    from f3dasm._src.optimization.agent_prompts import (
+    from f3dasm._src.agentic.agent_prompts import (
         CHECKPOINT_STRATEGIZER_PROMPT,
     )
 
@@ -653,7 +653,7 @@ def test_checkpoint_prompt_contains_comment_log():
 
 def test_implementer_reasoning_protocol_tag_appears_once():
     """IMPLEMENTER_SYSTEM_PROMPT has exactly one <reasoning_protocol> pair."""
-    from f3dasm._src.optimization.agent_prompts import (
+    from f3dasm._src.agentic.agent_prompts import (
         IMPLEMENTER_SYSTEM_PROMPT,
     )
 
@@ -666,7 +666,7 @@ def test_implementer_reasoning_protocol_tag_appears_once():
 
 def test_implementer_reasoning_protocol_content():
     """reasoning_protocol mentions Stages 1-3 and their headings."""
-    from f3dasm._src.optimization.agent_prompts import (
+    from f3dasm._src.agentic.agent_prompts import (
         IMPLEMENTER_SYSTEM_PROMPT,
     )
 
@@ -691,7 +691,7 @@ def test_implementer_reasoning_protocol_content():
 
 def test_no_domain_specific_leakage_extended():
     """No prompt contains 'coilable', 'sigma_crit', or 'Bessa'."""
-    from f3dasm._src.optimization.agent_prompts import (
+    from f3dasm._src.agentic.agent_prompts import (
         CHECKPOINT_STRATEGIZER_PROMPT,
         IMPLEMENTER_RESET_PROMPT_TEMPLATE,
         IMPLEMENTER_SYSTEM_PROMPT,
@@ -721,7 +721,7 @@ def test_no_domain_specific_leakage_extended():
 
 def test_run_paths_preamble_template_exists_and_is_string():
     """RUN_PATHS_PREAMBLE_TEMPLATE is a non-empty string."""
-    from f3dasm._src.optimization.agent_prompts import (
+    from f3dasm._src.agentic.agent_prompts import (
         RUN_PATHS_PREAMBLE_TEMPLATE,
     )
 
@@ -746,7 +746,7 @@ def test_run_paths_preamble_template_placeholders():
     ``strategizer_notes_dir`` as literal labels in its body so that
     the Strategizer understands the semantics of each path.
     """
-    from f3dasm._src.optimization.agent_prompts import (
+    from f3dasm._src.agentic.agent_prompts import (
         RUN_PATHS_PREAMBLE_TEMPLATE,
     )
 
@@ -780,7 +780,7 @@ def test_run_paths_preamble_template_placeholders():
 
 def test_workspace_preamble_template_exists_and_is_string():
     """WORKSPACE_PREAMBLE_TEMPLATE is a non-empty string."""
-    from f3dasm._src.optimization.agent_prompts import (
+    from f3dasm._src.agentic.agent_prompts import (
         WORKSPACE_PREAMBLE_TEMPLATE,
     )
 
@@ -805,7 +805,7 @@ def test_workspace_preamble_template_placeholder_and_no_tmp():
     the Implementer understands that scratch files outside the workspace
     will be lost.
     """
-    from f3dasm._src.optimization.agent_prompts import (
+    from f3dasm._src.agentic.agent_prompts import (
         WORKSPACE_PREAMBLE_TEMPLATE,
     )
 
@@ -834,7 +834,7 @@ def test_workspace_preamble_template_placeholder_and_no_tmp():
 
 def test_implementer_report_retry_prompt_exists_and_is_string():
     """IMPLEMENTER_REPORT_RETRY_PROMPT is a non-empty string."""
-    from f3dasm._src.optimization.agent_prompts import (
+    from f3dasm._src.agentic.agent_prompts import (
         IMPLEMENTER_REPORT_RETRY_PROMPT,
     )
 
@@ -858,7 +858,7 @@ def test_implementer_report_retry_prompt_required_headings():
     The runtime greps for ``## Report``; the four ``###`` subsections
     are structural requirements checked by ``_classify_failed_*``.
     """
-    from f3dasm._src.optimization.agent_prompts import (
+    from f3dasm._src.agentic.agent_prompts import (
         IMPLEMENTER_REPORT_RETRY_PROMPT,
     )
 
@@ -883,7 +883,7 @@ def test_implementer_report_retry_prompt_required_headings():
 def test_reflect_diagnosis_short_exists_and_keyword():
     """REFLECT_DIAGNOSIS_SHORT is a non-empty string mentioning 'unusually
     short'."""
-    from f3dasm._src.optimization.agent_prompts import (
+    from f3dasm._src.agentic.agent_prompts import (
         REFLECT_DIAGNOSIS_SHORT,
     )
 
@@ -905,7 +905,7 @@ def test_reflect_diagnosis_short_exists_and_keyword():
 def test_reflect_diagnosis_capability_limit_exists_and_keyword():
     """REFLECT_DIAGNOSIS_CAPABILITY_LIMIT is a non-empty string mentioning
     'capability'."""
-    from f3dasm._src.optimization.agent_prompts import (
+    from f3dasm._src.agentic.agent_prompts import (
         REFLECT_DIAGNOSIS_CAPABILITY_LIMIT,
     )
 
@@ -933,7 +933,7 @@ def test_reflect_diagnosis_missing_subsections_template():
     The literal substring ``{missing_subsections}`` must exist in the
     raw template text so the contract is visible to code readers.
     """
-    from f3dasm._src.optimization.agent_prompts import (
+    from f3dasm._src.agentic.agent_prompts import (
         REFLECT_DIAGNOSIS_MISSING_SUBSECTIONS_TEMPLATE,
     )
 
@@ -969,7 +969,7 @@ def test_reflect_diagnosis_missing_subsections_template():
 
 def test_reflect_diagnosis_no_report_heading_exists_and_keyword():
     """REFLECT_DIAGNOSIS_NO_REPORT_HEADING is a non-empty string."""
-    from f3dasm._src.optimization.agent_prompts import (
+    from f3dasm._src.agentic.agent_prompts import (
         REFLECT_DIAGNOSIS_NO_REPORT_HEADING,
     )
 
@@ -990,7 +990,7 @@ def test_reflect_diagnosis_no_report_heading_exists_and_keyword():
 
 def test_reflect_diagnosis_default_exists_and_keyword():
     """REFLECT_DIAGNOSIS_DEFAULT is a non-empty string."""
-    from f3dasm._src.optimization.agent_prompts import (
+    from f3dasm._src.agentic.agent_prompts import (
         REFLECT_DIAGNOSIS_DEFAULT,
     )
 
@@ -1019,7 +1019,7 @@ def test_runtime_imports_all_new_constants():
     import pathlib
 
     runtime_path = pathlib.Path(
-        "src/f3dasm/_src/optimization/agent_runtime.py"
+        "src/f3dasm/_src/agentic/agent_runtime.py"
     )
     if not runtime_path.is_absolute():
         repo_root = pathlib.Path(__file__).parent.parent.parent
